feat(context): consume confirmed events (#143)
* feat: implement robust  (#128)

* feat: implement tmrobust

* feat: apply robust client

* feat: implement robust grpc wrapper

* feat: add robust

* feat: change alias name

* apply code review

* feat(pkg): implemen txconv

* fix(proto): use uint64 instead

* feat(event): toss confirmed event to context

* fix(pkg): resolve import cycle

* proto: add signer

* env: mod

* proto: gen files

* fix: lint

* state(context): add signer repo

* proto: use unified chain type

* context: confirmed events

* make: add lint

* app: fix wrong network name

* app: use utils

---------

+++ x/event/keeper/poll_test.go
@@ -46,7 +46,7 @@ func mockEvent(t *testing.T) *types.Event {
 			Req: &types.TxReqEvent{
 				DestChain: Chains[rand.Int()%len(Chains)],
 				DestAddr:  bz,
-				OpId:      rand.Uint32(),
+				OpId:      rand.Uint64(),
 				OpArgs:    [][]byte{bz},
 				Funds: []*mitotypes.Coin{
 					{
